refactor(admin): adopt canonical normalizeObjectIdLike
