+++ apps/admin/src/lib/utils/idUtils.ts
@@ -1,11 +1,4 @@
-/**
- * Normalizes an ID that could be a string or a MongoDB-style object { id, _id }
- */
-export const normalizeObjectIdLike = (value: unknown): string => {
-    if (typeof value === "string") return value;
-    if (value && typeof value === "object") {
-        const record = value as { id?: string; _id?: string };
-        return record.id || record._id || "";
-    }
-    return "";
-};
+import { sanitizeMongoObjectId, normalizeObjectIdLike } from "@esparex/shared";
+
+export { sanitizeMongoObjectId, normalizeObjectIdLike };
+
